pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[project]
name = "photowebpage"
version = "0.1.0"
authors = [
    {name = "Tim Schaefer", email="ts+code@rcmd.org"}
]
description="Generate a simple, static HTML page with an image gallery that displays all images in a folder."
readme = "README.md"
keywords = ["webpage", "photos", "images", "HTML"]
license = {file = "LICENSE"}
classifiers=[
        "Programming Language :: Python :: 3 :: Only",
        "Programming Language :: Python :: 3.9",
        "Programming Language :: Python :: 3.10",
        "Programming Language :: Python :: 3.11",
        "Development Status :: 3 - Alpha",
        "Intended Audience :: Developers",
        "License :: OSI Approved :: MIT License",
        "Operating System :: POSIX :: Linux",
        "Operating System :: Microsoft :: Windows",
        "Operating System :: MacOS",
        "Natural Language :: English",
]

dependencies = ["pillow"]

[project.urls]
homepage = "https://github.com/dfsp-spirit/photowebpage"
documentation = "https://github.com/dfsp-spirit/photowebpage"
repository = "https://github.com/dfsp-spirit/photowebpage"

[project.scripts]
photogallery = "photowebpage.cli:main"
